Broaden exception catching for single-shot timers
Unfortunately, single-shot timers are poorly implemented on several matplotlib backends and there are platform-specific issues in many cases.  It's best to throw a dangerously-broad net, in this specific case.  It should always be safe to continue, even if single-shot timers don't work.  Worst case, timers don't work and we are a bit less efficient than we should be.