Project in screen space to a point within segment.
Projection of clicked point to line is now done in screen rather than
data units, which helps if the two axes have widely different scalings.
Additionally, it is also always within the bounds of the segment, which
was not necessarily the case before for quickly varying data.

For example, `plt.plot(np.random.rand(100))` is improved on both
accounts.