Keybindings kwarg only need updated keybindings
Currently, specifying `keybindings` as a dict that does not contain all
keybindings will raise an error.  This was unintentional, and causes the
addition of the "next" and "prev" keybindings to break backwards
compaibility.  This fixes the issue by updating a copy of the default
keybindings, which is what should have been done all along.